monitoramento_sensores_iot/requirements.txt
Faker==37.0.1
kafka-python==2.1.2